Don't wrap fields
Wrapping fields looks a bit nicer, but actually makes things quite a bit
harder to process with standard shell tools.

Want all errors with all data? With wrapping just a "grep" won't be
enough, and even grep -A1 won't be because not all might be wrapped.

This just makes everything easier at the expensive of being harder to
read on dev. I considered adding an option, but don't really want to add
the complexity. We'll see how this goes.